+++ grid_toplevel/grid_toplevel/grid/grid_ui.h
@@ -33,7 +33,7 @@ void grid_port_process_ui(struct grid_port* por);
 uint8_t grid_ui_model_init(struct grid_ui_model* mod, uint8_t len);
 
 
-uint8_t grid_ui_report_init(struct grid_ui_model* mod, uint8_t index, uint8_t* p, uint8_t p_len, uint8_t* h, uint8_t h_len);
+uint8_t grid_ui_report_init(struct grid_ui_model* mod, uint8_t index, uint8_t* p, uint32_t p_len, uint8_t* h, uint32_t h_len);
 
 uint8_t grid_ui_report_render(struct grid_ui_model* mod, uint8_t index, uint8_t* target);
